Combine module list tests
